Move types in a module
